Move Router class to its own file
Separated the Router class from src/index.tsx definition to its own file (src/Router.tsx)
