+++ src/index.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import { generatePath, Route  } from 'react-router-dom';
+import { Router } from './Router';
 
 import NamesakeLink from './NamesakeLink';
 import NamesakeRoute from './NamesakeRoute';
@@ -23,43 +24,4 @@ export interface INamesakeRouterState {
   transitionTo(routeName: string, params: {}, state: {}): void;
 }
 
-export class Router extends React.Component<INamesakeRouterProps, {}> {
-  private static defaultProps = {
-    routes: {},
-  };
-
-  public state = {
-    path: (routeName: string): string => {
-      const route = this.state.routes[routeName];
-      if (!route) {
-        throw new Error(`Unable to find route for '${routeName}'`);
-      }
-      return route;
-    },
-
-    routes: this.props.routes,
-
-    route: (routeName: string, params={}): string => {
-      const route = this.state.routes[routeName];
-      if (!route) {
-        throw new Error(`Unable to find route for '${routeName}'`);
-      }
-      return generatePath(route, params);
-    },
-
-    transitionTo: (routeName: string, params={}, state={}): void => {
-      this.props.push(this.state.route(routeName, params), state);
-    },
-  };
-
-  public render() {
-    const { children, routes, push, ...props } = this.props;
-    return (
-      <NamesakeProvider value={this.state}>
-        {children}
-      </NamesakeProvider>
-    );
-  }
-}
-
-export { NamesakeConsumer, NamesakeProvider, NamesakeLink, NamesakeRoute, WithNamesake };
+export { NamesakeConsumer, NamesakeProvider, NamesakeLink, NamesakeRoute,  Router, WithNamesake };
